Added groundwork to have clients aware of opponent disconnects
This is harder than it may seem. The clients have to have some
sort of method for checking the health of the server and making
sure the connection is still there. If the server has severed
the socket the client won't know until the next time it tries to
write or read to/from that socket. During board setup, the click
event is a good place for that.